refactor: excursion form
